fix: fix stale trial data caused by internal react re-route (#2763)
(cherry picked from commit 95db1fe458a34708d41ff40d9772fa750cffbf4a)
